+++ tests/unit/CRUD/base/base-batch-updating-test.js
@@ -1,14 +1,11 @@
-import { run } from '@ember/runloop';
-import { Promise } from 'rsvp';
-import { A } from '@ember/array';
-import { isNone } from '@ember/utils';
+import Ember from 'ember';
 import generateUniqueId from 'ember-flexberry-data/utils/generate-unique-id';
 
 export default function batchUpdating(store, assert) {
   assert.expect(14);
   let done = assert.async();
 
-  run(() => {
+  Ember.run(() => {
     initTestData(store)
 
       // Without relationships.
@@ -17,7 +14,7 @@ export default function batchUpdating(store, assert) {
         const user2Id = records.people[1];
         const user3Id = records.people[2];
         const suggTypeId = records.suggestionTypes[0];
-        return Promise.all([
+        return Ember.RSVP.Promise.all([
           store.findRecord('ember-flexberry-dummy-application-user', user1Id)
             .then((returned1Record) => {
               returned1Record.set('name', 'Updated name for User 1');
@@ -56,7 +53,7 @@ export default function batchUpdating(store, assert) {
             const record4 = recordsForBatch[3];
             const record5 = recordsForBatch[4];
             const record6 = recordsForBatch[5];
-            return store.adapterFor('application').batchUpdate(store, A([record4, record2, record3, record1, record5, record6])).then((result) => {
+            return store.batchUpdate(Ember.A([record4, record2, record3, record1, record5, record6])).then((result) => {
               assert.equal(result.length, 6);
 
               assert.notOk(result[0].get('hasDirtyAttributes'));
@@ -94,13 +91,12 @@ export default function batchUpdating(store, assert) {
             store.unloadAll();
             return store.findRecord('ember-flexberry-dummy-suggestion-type', suggTypeId)
               .then((suggTypeRecord) =>
-                assert.ok(isNone(suggTypeRecord.get('parent')), 'Relationship change')
+                assert.ok(Ember.isNone(suggTypeRecord.get('parent')), 'Relationship change')
               );
           })
           .then(() => records);
       })
       .catch((e) => {
-        // eslint-disable-next-line no-console
         console.log(e, e.message);
         throw e;
       })
@@ -116,7 +112,7 @@ function initTestData(store) {
 
     // Attrs for creating suggestion.
     .then((parentType) =>
-      Promise.all([
+      Ember.RSVP.Promise.all([
         store.createRecord('ember-flexberry-dummy-application-user', {
           name: 'Vasya',
           eMail: '1@mail.ru',
@@ -138,7 +134,7 @@ function initTestData(store) {
         }).save()
       ])
     ).then((createdRecords) =>
-      new Promise((resolve) =>
+      new Ember.RSVP.Promise((resolve) =>
         resolve({
           people: createdRecords.slice(0, 3).map(item => item.get('id')),
           suggestionTypes: createdRecords.slice(3, 4).map(item => item.get('id'))