+++ tests/unit/CRUD/offline/offline-association-batch-updating-test.js
@@ -1,0 +1,6 @@
+import executeTest from './execute-offline-test';
+import associationBatchUpdating from '../base/association-batch-updating-test';
+
+executeTest('associationBatchUpdating', (store, assert) => {
+  associationBatchUpdating(store, assert);
+});